+++ plugins/nf-schema/src/main/nextflow/validation/ValidationExtension.groovy
@@ -35,6 +35,7 @@ import nextflow.validation.exceptions.SchemaValidationException
 import nextflow.validation.help.HelpMessage
 import nextflow.validation.validators.JsonSchemaValidator
 import nextflow.validation.samplesheet.SamplesheetConverter
+import nextflow.validation.summary.SummaryCreator
 import nextflow.validation.parameters.ParameterValidator
 import static nextflow.validation.utils.Colors.logColors
 import static nextflow.validation.utils.Files.paramsLoad
@@ -163,7 +164,7 @@ class ValidationExtension extends PluginExtensionPoint {
         Map options = null
     ) {
         def Map params = initialiseExpectedParams(session.params)
-        def validator = new ParameterValidator(config)
+        def ParameterValidator validator = new ParameterValidator(config)
         validator.validateParametersMap(
             options,
             params,
@@ -212,105 +213,17 @@ Please contact the pipeline maintainer(s) if you see this warning as a user.
     // Groovy Map summarising parameters/workflow options used by the pipeline
     //
     @Function
-    public LinkedHashMap paramsSummaryMap(
+    public Map paramsSummaryMap(
         Map options = null,
         WorkflowMetadata workflow
         ) {
-        
-        def String schemaFilename = options?.containsKey('parameters_schema') ? options.parameters_schema as String : config.parametersSchema
-        def Map params = session.params
-        
-        // Get a selection of core Nextflow workflow options
-        def Map workflowSummary = [:]
-        if (workflow.revision) {
-            workflowSummary['revision'] = workflow.revision
-        }
-        workflowSummary['runName']      = workflow.runName
-        if (workflow.containerEngine) {
-            workflowSummary['containerEngine'] = workflow.containerEngine
-        }
-        if (workflow.container) {
-            workflowSummary['container'] = workflow.container
-        }
-
-        workflowSummary['launchDir']    = workflow.launchDir
-        workflowSummary['workDir']      = workflow.workDir
-        workflowSummary['projectDir']   = workflow.projectDir
-        workflowSummary['userName']     = workflow.userName
-        workflowSummary['profile']      = workflow.profile
-        workflowSummary['configFiles']  = workflow.configFiles ? workflow.configFiles.join(', ') : ''
-
-        // Get pipeline parameters defined in JSON Schema
-        def Map paramsSummary = [:]
-        def Map paramsMap = paramsLoad( Path.of(getSchemaPath(session.baseDir.toString(), schemaFilename)) )
-        for (group in paramsMap.keySet()) {
-            def Map groupSummary = getSummaryMapFromParams(params, paramsMap.get(group) as Map)
-            config.summary.hideParams.each { hideParam ->
-                def List<String> hideParamList = hideParam.tokenize(".") as List<String>
-                def Integer indexCounter = 0
-                def Map nestedSummary = groupSummary
-                if(hideParamList.size() >= 2 ) {
-                    hideParamList[0..-2].each { it ->
-                        nestedSummary = nestedSummary?.get(it, null)
-                    }
-                }
-                if(nestedSummary != null ) {
-                    nestedSummary.remove(hideParamList[-1])
-                }
-            }
-            paramsSummary.put(group, groupSummary)
-        }
-        paramsSummary.put('Core Nextflow options', workflowSummary)
-        return paramsSummary
-    }
-
-
-    //
-    // Create a summary map for the given parameters
-    //
-    private Map getSummaryMapFromParams(Map params, Map paramsSchema) {
-        def Map summary = [:]
-        for (String param in paramsSchema.keySet()) {
-            if (params.containsKey(param)) {
-                def Map schema = paramsSchema.get(param) as Map 
-                if (params.get(param) instanceof Map && schema.containsKey("properties")) {
-                    summary.put(param, getSummaryMapFromParams(params.get(param) as Map, schema.get("properties") as Map))
-                    continue
-                }
-                def String value = params.get(param)
-                def String defaultValue = schema.get("default")
-                def String type = schema.type
-                if (defaultValue != null) {
-                    if (type == 'string') {
-                        // TODO rework this in a more flexible way
-                        if (defaultValue.contains('$projectDir') || defaultValue.contains('${projectDir}')) {
-                            def sub_string = defaultValue.replace('\$projectDir', '')
-                            sub_string     = sub_string.replace('\${projectDir}', '')
-                            if (value.contains(sub_string)) {
-                                defaultValue = value
-                            }
-                        }
-                        if (defaultValue.contains('$params.outdir') || defaultValue.contains('${params.outdir}')) {
-                            def sub_string = defaultValue.replace('\$params.outdir', '')
-                            sub_string     = sub_string.replace('\${params.outdir}', '')
-                            if ("${params.outdir}${sub_string}" == value) {
-                                defaultValue = value
-                            }
-                        }
-                    }
-                }
-
-                // We have a default in the schema, and this isn't it
-                if (defaultValue != null && value != defaultValue) {
-                    summary.put(param, value)
-                }
-                // No default in the schema, and this isn't empty or false
-                else if (defaultValue == null && value != "" && value != null && value != false && value != 'false') {
-                    summary.put(param, value)
-                }
-            }
-        }
-        return summary
+        def SummaryCreator creator = new SummaryCreator(config)
+        return creator.createSummaryMap(
+            options,
+            workflow,
+            session.baseDir.toString(),
+            session.params
+        )
     }
 
     //
